+++ src/backgrounds/Background.jsx
@@ -1,11 +1,13 @@
-export default function Login_Signup() {
-  return (
-    <div className="fixed inset-0 -z-10">
-      <img
-        src="/login_signup.jpg"
-        alt="codefest background"
-        className="h-full w-full object-cover"
-      />
-    </div>
-  );
+export default function Background({ image_path }) {
+    return (
+        <>
+            <img
+                alt="codefest"
+                src={image_path}
+                width={0}
+                height={0}
+                className="absolute top-0 left-0 h-[100vh] lg:h-[110vh] xl:h-[125vh] w-fit md:w-[100vw] object-cover lg:-mt-[10vh] xl:-mt-[25vh] z-[-1] overflow-y-hidden"
+            />
+        </>
+    );
 }
