+++ src/backgrounds/Login_Signup.jsx
@@ -1,0 +1,11 @@
+export default function Login_Signup() {
+  return (
+    <div className="fixed inset-0 -z-10">
+      <img
+        src="/login_signup.jpg"
+        alt="codefest background"
+        className="h-full w-full object-cover"
+      />
+    </div>
+  );
+}
